Fixed compiler warnings around implicit conversions losing floating-point precision

+++ Classes/MDACMochiDevStyle.m
@@ -91,7 +91,7 @@
 
 - (UIColor *)listCellBackgroundColor
 {
-    return [UIColor colorWithWhite:94./255. alpha:1];
+    return [UIColor colorWithWhite:(CGFloat)(94./255.) alpha:1];
 }
 
 - (UIColor *)listCellTextColor
@@ -101,12 +101,12 @@
 
 - (UIColor *)listCellDetailTextColor
 {
-    return [UIColor colorWithWhite:1 alpha:0.7];
+    return [UIColor colorWithWhite:1 alpha:(CGFloat)0.7];
 }
 
 - (UIColor *)listCellShadowColor
 {
-    return [UIColor colorWithWhite:0 alpha:0.6];
+    return [UIColor colorWithWhite:0 alpha:(CGFloat)0.6];
 }
 
 - (CGSize)listCellShadowOffset
@@ -126,7 +126,7 @@
 
 - (UIColor *)iconCellShadowColor
 {
-    return [UIColor colorWithWhite:0 alpha:0.8];
+    return [UIColor colorWithWhite:0 alpha:(CGFloat)0.8];
 }
 
 - (CGSize)iconCellShadowOffset
@@ -146,7 +146,7 @@
 
 - (UIColor *)textCellShadowColor
 {
-    return [UIColor colorWithWhite:0 alpha:0.8];
+    return [UIColor colorWithWhite:0 alpha:(CGFloat)0.8];
 }
 
 - (CGSize)textCellShadowOffset
